Use explicit 64-bit type so this works in 32-bit builds

+++ day6/parts.cc
@@ -4,14 +4,14 @@
 #include <vector>
 #include <iterator>
 int main() {
-   unsigned long counts[9] = {};
+   uint64_t counts[9] = {};
    while (!std::cin.eof()) {
-      unsigned long i;
+      uint64_t i;
       char comma;
       std::cin >> i >> comma;
       counts[i]++;
    }
    for (size_t i = 0; i != 257; counts[(i + 7) % 9] += counts[i%9], ++i)
       if (i == 80 || i == 256)
-         std::cout << "day " << i << ": " << std::accumulate(counts, counts + 9, 0UL) << std::endl;
+         std::cout << "day " << i << ": " << std::accumulate(counts, counts + 9, uint64_t(0)) << std::endl;
 }
